ai_search_with_adi/function_app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
python-dotenv
azure-functions
openai
azure-storage-blob
pandas
azure-identity
openpyxl
regex
nltk==3.8.1
bs4
azure-search
azure-search-documents
azure-ai-documentintelligence
azure-ai-textanalytics
azure-ai-vision-imageanalysis
PyMuPDF
aiohttp
